# apps/api/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workforce/save": {
      "post": {
        "operationId": "WorkforceController_save",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkforceSaveDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkforceRecordDto"
                }
              }
            }
          }
        },
        "tags": [
          "Workforce"
        ]
      }
    },
...
    "/api/v1/workforce/history": {
      "post": {
        "operationId": "WorkforceController_history",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkforceHistoryRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkforceRevisionDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Workforce"
        ]
      }
    },
    "/api/v1/session/context": {
      "get": {
        "operationId": "PlatformController_context[0]",
        "parameters": [
          {
            "name": "X-IOP-Demo",
            "in": "header",
            "description": "Value 1 required on POST, together with an allowed Origin and local session cookie.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoContextDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid bounded input or unavailable reference.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Select a configured local demo user."
          },
          "403": {
            "description": "Scope, current grants, host or browser origin denied."
          },
          "409": {
            "description": "Revision changed or import capacity/recovery conflict."
          },
          "422": {
            "description": "Exact analytical total exceeds the supported integer range."
          },
          "503": {
            "description": "Local mode inactive, maintenance, busy upload or persistence unavailable."
          }
        },
        "tags": [
          "Local analytical POC"
        ]
      }
    },
    "/api/v1/demo/context": {
      "get": {
        "operationId": "PlatformController_context[1]",
        "parameters": [
          {
            "name": "X-IOP-Demo",
            "in": "header",
            "description": "Value 1 required on POST, together with an allowed Origin and local session cookie.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoContextDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid bounded input or unavailable reference.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Select a configured local demo user."
          },
          "403": {
            "description": "Scope, current grants, host or browser origin denied."
          },
          "409": {
            "description": "Revision changed or import capacity/recovery conflict."
          },
          "422": {
            "description": "Exact analytical total exceeds the supported integer range."
          },
          "503": {
            "description": "Local mode inactive, maintenance, busy upload or persistence unavailable."
          }
        },
        "tags": [
          "Local analytical POC"
        ]
      }
    },
    "/api/v1/demo/user": {
      "post": {
        "operationId": "PlatformController_switchUser",
        "parameters": [
          {
            "name": "X-IOP-Demo",
            "in": "header",
            "description": "Value 1 required on POST, together with an allowed Origin and local session cookie.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoContextDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid bounded input or unavailable reference.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Select a configured local demo user."
          },
          "403": {
            "description": "Scope, current grants, host or browser origin denied."
          },
          "409": {
            "description": "Revision changed or import capacity/recovery conflict."
          },
          "422": {
            "description": "Exact analytical total exceeds the supported integer range."
          },
          "503": {
            "description": "Local mode inactive, maintenance, busy upload or persistence unavailable."
          }
        },
        "tags": [
          "Local analytical POC"
        ]
      }
    },
    "/api/v1/imports": {
      "get": {
        "operationId": "PlatformController_history",
        "parameters": [
          {
            "name": "X-IOP-Demo",
            "in": "header",
            "description": "Value 1 required on POST, together with an allowed Origin and local session cookie.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImportSummaryDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid bounded input or unavailable reference.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Select a configured local demo user."
          },
          "403": {
            "description": "Scope, current grants, host or browser origin denied."
          },
          "409": {
            "description": "Revision changed or import capacity/recovery conflict."
          },
          "422": {
            "description": "Exact analytical total exceeds the supported integer range."
          },
          "503": {
            "description": "Local mode inactive, maintenance, busy upload or persistence unavailable."
          }
        },
        "tags": [
          "Local analytical POC"
        ]
      },
      "post": {
        "operationId": "PlatformController_submit",
        "parameters": [
          {
            "name": "X-IOP-Demo",
            "in": "header",
            "description": "Value 1 required on POST, together with an allowed Origin and local session cookie.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSV-Filename",
            "in": "header",
            "description": "Supported Hitliste-YYYYMMDD.csv reporting label.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportReviewDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid bounded input or unavailable reference.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Select a configured local demo user."
          },
          "403": {
            "description": "Scope, current grants, host or browser origin denied."
          },
          "409": {
            "description": "Revision changed or import capacity/recovery conflict."
          },
          "422": {
            "description": "Exact analytical total exceeds the supported integer range."
          },
          "503": {
            "description": "Local mode inactive, maintenance, busy upload or persistence unavailable."
          }
        },
        "tags": [
          "Local analytical POC"
        ]
      }
    },
    "/api/v1/imports/{id}": {
      "get": {
        "operationId": "PlatformController_review",
        "parameters": [
          {
            "name": "X-IOP-Demo",
            "in": "header",
            "description": "Value 1 required on POST, together with an allowed Origin and local session cookie.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportReviewDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid bounded input or unavailable reference.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Select a configured local demo user."
          },
          "403": {
            "description": "Scope, current grants, host or browser origin denied."
          },
          "409": {
            "description": "Revision changed or import capacity/recovery conflict."
          },
          "422": {
            "description": "Exact analytical total exceeds the supported integer range."
          },
          "503": {
            "description": "Local mode inactive, maintenance, busy upload or persistence unavailable."
          }
        },
        "tags": [
          "Local analytical POC"
        ]
      }
    },
    "/api/v1/imports/{id}/recover": {
      "post": {
        "operationId": "PlatformController_recover",
        "parameters": [
          {
            "name": "X-IOP-Demo",
            "in": "header",
            "description": "Value 1 required on POST, together with an allowed Origin and local session cookie.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportReviewDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid bounded input or unavailable reference.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Select a configured local demo user."
          },
          "403": {
            "description": "Scope, current grants, host or browser origin denied."
          },
          "409": {
            "description": "Revision changed or import capacity/recovery conflict."
          },
          "422": {
            "description": "Exact analytical total exceeds the supported integer range."
          },
          "503": {
            "description": "Local mode inactive, maintenance, busy upload or persistence unavailable."
          }
        },
        "tags": [
          "Local analytical POC"
        ]
      }
    },
    "/api/v1/imports/{id}/original": {
      "get": {
        "operationId": "PlatformController_original",
        "parameters": [
          {
            "name": "X-IOP-Demo",
            "in": "header",
            "description": "Value 1 required on POST, together with an allowed Origin and local session cookie.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Original CSV bytes; requires imports.review.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid bounded input or unavailable reference.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Select a configured local demo user."
          },
          "403": {
            "description": "Scope, current grants, host or browser origin denied."
          },
          "409": {
            "description": "Revision changed or import capacity/recovery conflict."
          },
          "422": {
            "description": "Exact analytical total exceeds the supported integer range."
          },
          "503": {
            "description": "Local mode inactive, maintenance, busy upload or persistence unavailable."
          }
        },
        "tags": [
          "Local analytical POC"
        ]
      }
    },
    "/api/v1/analytics/availability": {
      "get": {
        "operationId": "PlatformController_availability",
        "parameters": [
          {
            "name": "X-IOP-Demo",
            "in": "header",
            "description": "Value 1 required on POST, together with an allowed Origin and local session cookie.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid bounded input or unavailable reference.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Select a configured local demo user."
          },
          "403": {
            "description": "Scope, current grants, host or browser origin denied."
          },
          "409": {
            "description": "Revision changed or import capacity/recovery conflict."
          },
          "422": {
            "description": "Exact analytical total exceeds the supported integer range."
          },
          "503": {
            "description": "Local mode inactive, maintenance, busy upload or persistence unavailable."
          }
        },
        "tags": [
          "Local analytical POC"
        ]
      }
    },
...
    "/api/v1/auth/password": {
      "post": {
        "operationId": "AccessController_password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDto"
                }
              }
            }
          }
        },
        "tags": [
          "Local access"
        ]
      }
    },
...
    "/api/v1/handover/query": {
      "post": {
        "operationId": "HandoverController_query",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoverSelectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoverPageDto"
                }
              }
            }
          }
        },
        "tags": [
          "Shift Handover"
        ]
      }
    },
...
  },
  "info": {
    "title": "IOP local API host",
    "description": "Health plus explicitly activated local analytical POC. Business operations require a configured local demo session, current site permissions and exact configured scope; no shared-user authentication claim.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "WorkforceSaveDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "settings",
              "worker",
              "schedule",
              "assignment"
            ]
          },
          "expectedRevision": {
            "type": "number"
          },
          "deleted": {
            "type": "boolean"
          },
          "data": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "kind",
          "expectedRevision",
          "deleted",
          "data"
        ]
      },
...
      "WorkforceHistoryRequestDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "settings",
              "worker",
              "schedule",
              "assignment"
            ]
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "kind",
          "id"
        ]
      },
...
      "SwitchUserDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "userId"
        ]
      },
...
    }
  }
}